added background music track and also started to balance generators

+++ generator.py
@@ -10,11 +10,11 @@ class Generator:
         self.growth_rate = growth_rate
         self.starting_rev = starting_rev
         self.owned = 0
-        self.cost = int(self.starting_cost)
+        self.cost = self.starting_cost
         self.production = 0
 
     def upgrade(self):
         self.owned += 1
-        self.cost = int(float(self.starting_cost) * math.pow(float(self.growth_rate), float(self.owned)))
-        self.production = int(float(self.starting_product) * float(self.owned) * float(self.starting_rev))
+        self.cost = float(self.starting_cost) * math.pow(float(self.growth_rate), float(self.owned))
+        self.production = float(self.starting_product) * float(self.owned) * float(self.starting_rev)
